requirements.txt
# Core dependencies
streamlit>=1.28.0
python-dotenv>=1.0.0

# Document processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Embeddings and vector store
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
chromadb>=0.4.0

# LLM
openai>=1.0.0
tiktoken>=0.5.0

# Utilities
pyyaml>=6.0
numpy>=1.24.0
pandas>=2.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
